refactor: recover docs and demo
